Test Connect to LocalDB, Added Classes for DB Data
Since this is a project with the intended goal of practicing more with SQL and working with SQL Server, I'm including the appSettings.json file here for transparency. No actual secret information will be committed to it. Normally I would include this in my gitignore, but I feel having it visible in this project can help with context.